requirements.txt
flask
python-dotenv
paho-mqtt
influxdb3-python
pandas
certifi